Refactor:recycle:: #54 Add extension hideKeyboard
